fixed exitCommunity autons
