+++ src/main/java/com/team766/robot/procedures/OnePieceExitCommunity.java
@@ -2,26 +2,31 @@ package com.team766.robot.procedures;
 
 import com.team766.framework.Context;
 import com.team766.framework.Procedure;
+import com.team766.odometry.PointDir;
 import com.team766.robot.Robot;
-import com.team766.robot.constants.FollowPointsInputConstants;
-import java.util.function.BooleanSupplier;
+
 import edu.wpi.first.wpilibj.DriverStation;
-import com.team766.odometry.*;
 
 public class OnePieceExitCommunity extends Procedure {
-
-	private static final double DIST = 4;
-
-	public void run(Context context) {
+	public void run (Context context) {
 		context.takeOwnership(Robot.drive);
 		context.takeOwnership(Robot.intake);
-		double startX = Robot.drive.getCurrentPosition().getX();
+		context.takeOwnership(Robot.gyro);
 		Robot.gyro.resetGyro180();
 		Robot.drive.setGyro(Robot.gyro.getGyroYaw());
-		new ReverseIntake().run(context);
-		Robot.drive.swerveDrive(0, FollowPointsInputConstants.SPEED, 0);
-		context.waitFor(() -> Math.abs(Robot.drive.getCurrentPosition().getX() - startX) > DIST);
-		Robot.drive.stopDriveMotors();
-		Robot.drive.stopSteerMotors();
+		switch (DriverStation.getAlliance()) {
+			case Blue:
+				Robot.drive.setCurrentPosition(new PointDir(2, 0.75));
+				break;
+			case Red:
+				Robot.drive.setCurrentPosition(new PointDir(14.5, 0.75));
+				break;
+			case Invalid: //drop down
+			default: 
+				log("invalid alliance");
+				return;
+		}
+		log("exiting");
+		new OPECHelper().run(context);
 	}
 }
